+++ main.go
@@ -41,7 +41,7 @@ var (
 	tlsKeyFile  string
 
 	dryRun     bool
-	configPath = "config.yml"
+	configPath string
 
 	defaultConfig = Config{
 		CacheTimeIndex:      10 * time.Second,
@@ -55,10 +55,11 @@ func init() {
 	flag.StringVar(&address, "address", "localhost:8080", "Address of the package-registry service.")
 	flag.StringVar(&tlsCertFile, "tls-cert", "", "Path of the TLS certificate.")
 	flag.StringVar(&tlsKeyFile, "tls-key", "", "Path of the TLS key.")
+	flag.StringVar(&configPath, "config", "config.yml", "Path to the configuration file.")
 	flag.StringVar(&httpProfAddress, "httpprof", "", "Enable HTTP profiler listening on the given address.")
 	// This flag is experimental and might be removed in the future or renamed
-	flag.BoolVar(&dryRun, "dry-run", false, "Runs a dry-run of the registry without starting the web service (experimental)")
-	flag.BoolVar(&packages.ValidationDisabled, "disable-package-validation", false, "Disable package content validation")
+	flag.BoolVar(&dryRun, "dry-run", false, "Runs a dry-run of the registry without starting the web service (experimental).")
+	flag.BoolVar(&packages.ValidationDisabled, "disable-package-validation", false, "Disable package content validation.")
 }
 
 type Config struct {
@@ -70,7 +71,7 @@ type Config struct {
 }
 
 func main() {
-	flag.Parse()
+	parseFlags()
 	log.Println("Package registry started.")
 	defer log.Println("Package registry stopped.")